modify allBoards api, now with thunk
