+++ react/src/former-code/canvas-draggable.tsx
@@ -1,6 +1,6 @@
 //draggable https://www.npmjs.com/package/react-draggable#react-draggable
 import { useState, useRef, useEffect } from "react";
-import { Card } from '../app/fetch-data/apiSlice';
+import { Card } from '../app/fetch-data/apiSlice old';
 import './card.css';
 
 import Draggable, { DraggableData, DraggableEvent } from 'react-draggable';
