+++ react/src/former-code/canvas-svg.tsx
@@ -1,7 +1,7 @@
 //dragNdrop svg: https://gist.github.com/hashrock/0e8f10d9a233127c5e33b09ca6883ff4
 
 import React, { useState } from 'react';
-import { Card } from '../app/fetch-data/apiSlice';
+import { Card } from '../app/fetch-data/apiSlice old';
 
 import './card.css';
