+++ react/src/pages/single-board/elements/arrowFocus.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 
-import { Arrow } from '../../../app/fetch-data/apiSlice';
+import { Arrow } from '../../../app/fetch-data/dataTypes';
 import "./arrow.css"
 
 //we need that to read the state
